# contracts/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API behavior, compatibility rules, and endpoint contract requirements.",
    "version": "1.25.1"
  },
  "paths": {
...
    "/api/spool/{namespace}/{query_id}.parquet": {
      "get": {
        "summary": "GET /api/spool/{namespace}/{query_id}.parquet",
        "responses": {
          "200": {
            "description": "Contract response: application/octet-stream (parquet)"
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "410": {
            "description": "Error response (see contract error format)"
          }
        },
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-cdd-response-contract": "application/octet-stream (parquet)",
        "x-cdd-errors": "400/410"
      }
    },
...
    "/api/downtime-analysis/query": {
      "post": {
        "summary": "POST /api/downtime-analysis/query",
        "responses": {
          "201": {
            "description": "Contract response: DowntimeQueryResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DowntimeQueryResponse"
                }
              }
            }
          },
          "202": {
            "description": "Error response (see contract error format)"
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "500": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Contract request: JSON body (schema not machine-resolved; see contract)",
          "content": {},
          "x-cdd-unresolved": true
        },
        "x-cdd-errors": "202/400/500"
      }
    },
    "/api/downtime-analysis/options": {
      "get": {
        "summary": "GET /api/downtime-analysis/options",
        "responses": {
          "200": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-cdd-errors": "500"
      }
    },
    "/api/downtime-analysis/view": {
      "get": {
        "summary": "GET /api/downtime-analysis/view",
        "responses": {
          "200": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "410": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-cdd-errors": "400/410"
      }
    },
    "/api/downtime-analysis/equipment-detail": {
      "get": {
        "summary": "GET /api/downtime-analysis/equipment-detail",
        "responses": {
          "200": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "410": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-cdd-errors": "400/410"
      }
    },
    "/api/downtime-analysis/event-detail": {
      "get": {
        "summary": "GET /api/downtime-analysis/event-detail",
        "responses": {
          "200": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "410": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-cdd-errors": "400/410"
      }
    },
    "/api/downtime-analysis/export-equipment-detail": {
      "get": {
        "summary": "GET /api/downtime-analysis/export-equipment-detail",
        "responses": {
          "200": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "410": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-cdd-errors": "400/410"
      }
    },
    "/api/downtime-analysis/export-event-detail": {
      "get": {
        "summary": "GET /api/downtime-analysis/export-event-detail",
        "responses": {
          "200": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "410": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-cdd-errors": "400/410"
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "GenericSuccessResponse": {
        "type": [
          "object",
          "null"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {},
          "error": {},
          "meta": {
            "type": "object",
            "properties": {
              "timestamp": {
                "type": "string"
              },
              "app_version": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "DowntimeQueryResponse": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "error": {},
          "data": {
            "type": "object",
            "properties": {
              "query_id": {
                "type": "string"
              },
              "base_spool_url": {
                "type": "string"
              },
              "jobs_spool_url": {
                "type": "string"
              },
              "taxonomy": {
                "type": "object"
              },
              "summary": {},
              "daily_trend": {
                "type": "array"
              },
              "big_category": {
                "type": "array"
              },
              "top_reasons": {
                "type": "array"
              }
            }
          },
          "meta": {
            "type": "object",
            "required": [
              "timestamp"
            ],
            "properties": {
              "timestamp": {
                "type": "string"
              },
              "app_version": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  }
}